Running tests.
