+++ .circleci/config.yml
@@ -1,0 +1,38 @@
+version: 2
+jobs:
+  build:
+    parallelism: 1
+    machine:
+      image: circleci/classic:latest
+    environment:
+      TEST_RESULTS: /tmp/test-results
+    steps:
+      - checkout
+      - run:
+          name: Setup composer config
+          command: |
+            cd ~
+            mkdir .composer
+            cd .composer
+            bash -c 'echo "{\"github-oauth\": {\"github.com\": \"$GITHUB_TOKEN\"}}"' > auth.json
+      - run:
+          name: Install DKTL
+          command: |
+            cd ~
+            chmod 777 ./project/bin/dktl
+            export PATH=$PATH:~/project/bin
+            which dktl
+      - run:
+          name: Initialize Project
+          command: |
+            export PATH=$PATH:~/project/bin
+            cd ~/project
+            dktl init
+      - run:
+          name: Run tests
+          command: |
+            export PATH=$PATH:~/project/bin
+            cd ~/project
+            dktl init
+            dktl dc exec cli ./vendor/bin/phpunit
+
