+++ src/Command/BasicCommands.php
@@ -2,8 +2,8 @@
 
 namespace DkanTools\Command;
 
-use Robo\Result;
 use DkanTools\Util\Util;
+use Symfony\Component\Filesystem\Filesystem;
 
 /**
  * This is project's console commands configuration for Robo task runner.
@@ -12,155 +12,281 @@ use DkanTools\Util\Util;
  */
 class BasicCommands extends \Robo\Tasks
 {
+    const DRUPAL_FOLDER_NAME = "docroot";
+
     /**
-     * Run drush command on current site.
+     * Get all necessary dependencies and "make" a working codebase.
      *
-     * Run drush command on current site. For instance, to clear caches, run
-     * "dktl drush cc all".
+     * Running `dktl make` will:
+     *   1. Modify the stock drupal composer.json file to merge in anything in src/make
+     *   2. Use composer to download and build all php dependencies.
+     *   3. Symlink a number of dirs from /src into docroot.
+     *   4. If requested, pull the DKAN frontend application into docroot.
      *
-     * @param array $cmd Array of arguments to create a full Drush command.
+     * @option yes
+     *   Skip confirmation step, overwrite existing no matter what. Use with caution!
+     * @option prefer-dist
+     *   Prefer dist for packages. See composer docs.
+     * @option prefer-source
+     *   Prefer dist for packages. See composer docs.
+     * @option no-dev
+     *   Skip installing packages listed in require-dev.
+     * @option optimize-autoloader
+     *   Convert PSR-0/4 autoloading to classmap to get a faster autoloader.
+     * @option frontend
+     *   Build with the DKAN frontend application.
+     * @option tag
+     *   Specify DKAN tagged release to build.
+     * @option branch
+     *   Specify DKAN branch to build.
      */
-    public function drush(array $cmd)
+    public function make($opts = [
+        'yes|y' => false,
+        'prefer-source' => false,
+        'prefer-dist' => false,
+        'no-dev' => false,
+        'optimize-autoloader' => false,
+        'frontend' => false,
+        'tag' => null,
+        'branch' => null,
+        ])
     {
-        $drupal_root = Util::getProjectDocroot();
-        $drushExec = $this->taskExec('drush')->dir($drupal_root);
-        foreach ($cmd as $arg) {
-            $drushExec->arg($arg);
+        $this->io()->section("Running dktl make");
+
+        // Add project dependencies.
+        $this->addDrush();
+        $this->addDkan2($opts);
+
+        // Run composer install while passing the options.
+        $composerBools = ['prefer-source', 'prefer-dist', 'no-dev', 'optimize-autoloader'];
+        $install = $this->taskComposerInstall();
+        foreach ($composerBools as $composerBool) {
+            if ($opts[$composerBool] === true) {
+                $install->option($composerBool);
+            }
         }
-        return $drushExec->run();
+        $install->run();
+
+        // Symlink dirs from src into docroot.
+        $this->addSymlinksToDrupalRoot();
+
+        if ($opts['frontend'] === true) {
+            $this->installFrontend();
+        }
+
+        $this->io()->success("dktl make completed.");
     }
 
-    /**
-     * Proxy to composer.
-     */
-    public function composer(array $cmd)
+    private function addDrush()
     {
-        $exec = $this->taskExec('composer');
-        foreach ($cmd as $arg) {
-            $exec->arg($arg);
+        $addDrush = $this->taskComposerRequire()
+          ->dependency("drush/drush")
+          ->run();
+        if ($addDrush->getExitCode() != 0) {
+            $this->io()->error('Unable to add Drush as a project dependencies.');
+            exit;
         }
-        return $exec->run();
+        $this->io()->success("drush/drush added as a project dependency.");
     }
 
-
-    /**
-     * Performs common tasks when switching databases or code bases.
-     *
-     * Operations like running rr and updb. It also runs environment
-     * switching which is provided by the environment module.
-     *
-     * @param string $target_environment
-     *   One of the site environments. DKTL provides 4 by default: local,
-     *   development, test, and production.
-     */
-    public function deploy($target_environment)
+    private function addDkan2(array $opts)
     {
-        $project = Util::getProjectDirectory();
-        $script = "{$project}/src/script/deploy.sh";
-        $docroot = Util::getProjectDocroot();
-
-        if (file_exists($script)) {
-            $command = "{$script} {$docroot} {$target_environment}";
-            $this->_exec($command);
-        }
-    }
-
-    /**
-     * Check for existence of XDEBUG_DKTL environment variable.
-     */
-    private function xdebugCheck()
-    {
-        $xdebugDktl = getenv("XDEBUG_DKTL");
-        if (!$xdebugDktl) {
-            throw new \Exception("XDEBUG_DKTL environment variable must be "
-            . " set to use this command.");
-        }
-    }
-
-    /**
-     * Start xdebug on CLI and web containers.
-     *
-     * This command adds a .ini file to your src/docker/etc/php directory and
-     * restarts the CLI and web containers. It checks for compatibility by
-     * looking for a DKTL_XDEBUG environment variable. This is set by default in
-     * DKAN Tools' containers, but you may need to set it again if using your
-     * own.
-     *
-     * Adding /src/docker/etc to your project .gitignore is reccomended.
-     */
-    public function xdebugStart()
-    {
-        $this->xdebugCheck();
-        
-        $platform = getenv("PLATFORM");
-        $sourceFile = ($platform == 'Darwin') ? 'xdebug-macos.ini' : 'xdebug-linux.ini';
-        $dktlRoot = Util::getDktlDirectory();
-        $this->io()->text("Creating new xdebug.ini file for {$platform} platform.");
-        
-        $f = 'src/docker/etc/php/xdebug.ini';
-        if (file_exists($f)) {
-            throw new \Exception("File {$f} already exists.");
+        // Find Dkan2 version from options' tag or branch values.
+        if ($opts['tag']) {
+            $dkanVersion = $opts['tag'];
+        } elseif ($opts['branch']) {
+            $dkanVersion = "dev-{$opts['branch']}";
         }
 
-        $result = $this->taskWriteToFile($f)
-            ->textFromFile("$dktlRoot/assets/docker/etc/php/$sourceFile")
+        $addDkan2 = $this->taskComposerRequire()
+            ->dependency("getdkan/dkan2", $dkanVersion)
             ->run();
-
-        Util::directoryAndFileCreationCheck($result, $f, $this->io());
+        if ($addDkan2->getExitCode() != 0) {
+            $this->io()->error('Unable to add Drush and Dkan2 dependencies.');
+            exit;
+        }
+        $this->io()->success("getdkan/dkan2 added as a project dependency.");
     }
 
-    /**
-     * Stop xdebug on CLI and web containers.
-     *
-     * Removes the xdebug.ini file and restarts CLI and web containers. See
-     * xdebug:start for more information.
-     */
-    public function xdebugStop()
+    private function addSymlinksToDrupalRoot()
     {
-        $this->xdebugCheck();
+        $targetsAndLinks = [
+            ['target' => 'src/site',    'link' => '/sites/default'],
+            ['target' => 'src/modules', 'link' => '/modules/custom'],
+            ['target' => 'src/themes',  'link' => '/themes/custom'],
+            ['target' => 'src/schema',  'link' => '/schema'],
+        ];
+        foreach ($targetsAndLinks as $targetAndLink) {
+            $this->docrootSymlink(
+                $targetAndLink['target'],
+                self::DRUPAL_FOLDER_NAME . $targetAndLink['link']
+            );
+        }
+    }
 
-        $f = 'src/docker/etc/php/xdebug.ini';
-        $result = unlink($f);
-        if ($result) {
-            $this->io()->success("Removed xdebug.ini; restarting.");
+    private function installFrontend()
+    {
+        $this->io()->section('Adding frontend application');
+
+        $result = $this->downloadFrontend();
+
+        if ($result && $result->getExitCode() === 0) {
+            $this->io()->note(
+                'Successfully downloaded data-catalog-frontend to /src/frontend'
+            );
+        }
+
+        if (file_exists('src/frontend')) {
+            $this->docrootSymlink('src/frontend', self::DRUPAL_FOLDER_NAME . '/data-catalog-frontend');
+        }
+
+        $this->io()->note(
+            'You are building DKAN with the React frontend application. ' .
+            'In order for the frontend to find the correct routes to work correctly,' .
+            'you will need to enable the dkan_frontend module . ' .
+            'Do this by running "dktl install" with the "--frontend" option as well, ' .
+            'or else run "drush en dkan_frontend" after installation.'
+        );
+    }
+
+    public function installphpunit()
+    {
+        $result = $this->taskExec("which phpunit")->run();
+        if ($result->getExitCode() == 0) {
+            $this->io()->text('phpunit is already installed.');
+            return true;
+        }
+
+        $this->io()->text('Installing phpunit.');
+
+        $result = $this->taskFilesystemStack()->stopOnFail()
+            ->remove('/root/.composer/vendor')
+            ->remove('/root/.composer/composer.lock')
+            ->run();
+        if ($result->getExitCode() != 0) {
+            $this->io()->error('could not remove root composer files.');
             return $result;
-        } else {
-            throw new \Exception("Failed, xdebug.ini not found.");
         }
+        $result = $this->taskComposerRequire()
+            ->dependency('phpunit/phpunit', "7.5.18")
+            ->dir('/root/.composer')
+            ->run();
+
+        $result = $this->taskExec("ln -s /root/.composer/vendor/bin/phpunit /usr/local/bin/phpunit")->run();
+
+        return $result;
     }
 
-  /**
-   * Create a database dump excluding devel and datastore tables.
-   *
-   * Run drush command on $alias to create a database dump excluding tables
-   * related to devel and datastore.
-   *
-   * @param String $alias Drush alias of the site we want the db from.
-   */
-    public function siteGrabDatabase($alias)
+    /**
+     * Link src/modules to docroot/sites/all/modules/custom.
+     */
+    private function docrootSymlink($target, $link)
     {
-      // Tables for which we want the structure and not the data.
-        $structure_tables_common = array(
-        'accesslog', 'batch', 'cache', 'cache_*', '*_cache', 'ctools_views_cache',
-        'ctools_object_cache', 'flood', 'history', 'queue', 'search_*',
-        'semaphore', 'sessions', 'watchdog'
-        );
-        $structure_tables_devel = array('devel_queries', 'devel_times');
-        $structure_tables_webform = array('webform_submitted_data');
-        $structure_tables = array_merge(
-            $structure_tables_common,
-            $structure_tables_devel,
-            $structure_tables_webform
-        );
-        $structure_tables_list = implode(', ', $structure_tables);
-      // Tables to be completely skipped.
-        $skip_tables = array('dkan_datastore_*');
-        $skip_tables_list = implode(', ', $skip_tables);
+        $project_dir = Util::getProjectDirectory();
+        $target = $project_dir . "/{$target}";
+        $link = $project_dir . "/{$link}";
+        $link_parts = pathinfo($link);
+        $link_dirname = $link_parts['dirname'];
+        $target_path_relative_to_link = (new Filesystem())->makePathRelative($target, $link_dirname);
 
-        return $this->taskExec("drush $alias sql-dump")
-            ->option('structure-tables-list', $structure_tables_list)
-            ->option('skip-tables-list', $skip_tables_list)
-            ->rawArg('> excluded_tables.sql')
+        if (!file_exists($target) || !file_exists(self::DRUPAL_FOLDER_NAME)) {
+            $this->io()->warning(
+                "Skipping linking $target. Folders $target and '" .
+                self::DRUPAL_FOLDER_NAME."' must both be present to create link."
+            );
+            return;
+        }
+
+        $result = $this->taskFilesystemStack()->stopOnFail()
+            ->remove($link)
+            ->symlink($target_path_relative_to_link, $link)
             ->run();
+
+        if ($result->getExitCode() != 0) {
+            $this->io()->warning('Could not create link');
+        } else {
+            $this->io()->success("Symlinked $target to $link");
+        }
+        return $result;
+    }
+
+    private function downloadFrontend()
+    {
+        $result = $this->taskExec('git clone')
+            ->option('depth', '1')
+            ->option('branch', 'master')
+            ->arg('https://github.com/GetDKAN/data-catalog-frontend.git')
+            ->arg('frontend')
+            ->dir('src')
+            ->run();
+        if ($result->getExitCode() != 0) {
+            $this->io()->error('Could not download front-end app.');
+            return $result;
+        }
+        $result = $this->_deleteDir('src/frontend/.git');
+        if ($result->getExitCode() != 0) {
+            $this->io()->error('Could not remove front-end git folder.');
+            return $result;
+        }
+
+        $this->io()->success('frontend application added.');
+        return $result;
+    }
+
+    public function install($opts = ['frontend' => false])
+    {
+        $result = $this->taskExec('drush si -y')
+            ->dir(Util::getProjectDocroot())
+            ->run();
+
+        if ($opts['frontend'] === true) {
+            $result = $this->taskExec('drush en -y')
+                ->arg('dkan_frontend')
+                ->dir(Util::getProjectDocroot())
+                ->run();
+        }
+        return $result;
+    }
+
+    /**
+     * Proxy to the phpunit binary.
+     *
+     * @param array $args  Arguments to append to full phpunit command.
+     */
+    public function phpunit(array $args)
+    {
+
+        $proj_dir = Util::getProjectDirectory();
+
+        $phpunit_executable = "{$proj_dir}/docroot/vendor/bin/phpunit";
+
+        $phpunitExec = $this->taskExec($phpunit_executable);
+
+        foreach ($args as $arg) {
+            $phpunitExec->arg($arg);
+        }
+
+        return $phpunitExec->run();
+    }
+
+    private function checkDrushCompatibility($version = '9')
+    {
+        if (version_compare($this->getDrushVersion(), $version) >= 0) {
+            return true;
+        }
+        return false;
+    }
+
+    private function getDrushVersion()
+    {
+        $result = $this->taskExec('drush --version')
+            ->dir(Util::getProjectDocroot())
+            ->printOutput(false)
+            ->run();
+        preg_match('/.+?(\d+\.\d+\.\d+)/', $result->getMessage(), $matches);
+        if (!isset($matches[1])) {
+            throw new \Exception("Could not determine Drush version on this system.");
+        }
+        return $matches[1];
     }
 }
